reuse logging code for both run listeners
junit 4 and 5
